+++ README.md
@@ -1,10 +1,10 @@
-Яндекс практикум. Проектная работа 9. Интерактивная страница с загрузкой/выгрузкой данных на сервер.
+Яндекс практикум. Проектная работа 11. Работа с модулями и Вэбпаками
 =============================================================
 Блок обучения JavaScript на Яндекс Практикуме.
 ---
 Актуальная версия проекта - v.0.0.1
 ---
-Описание: В проекте можно добавлять/удалять карточки городов, менять имя пользователя, его профессию и его аватар, ставить лайки и видеть лайки других пользователей. При заполнении форм добавления/изменения поля валидируются. Все данные подгружаются с сервера, при изменении и добавлении все данные сохраняются на сервере.
+Описание: В проекте код разбит на модули, собран Webpack-ом и выложен на Github
 ---
 Проект - домашнее задание в программе обучения Яндекс.Практикума.
 ---
