+++ webpack.config.js
@@ -1,0 +1,59 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WebpackMd5Hash = require('webpack-md5-hash');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // добавили плагин
+const webpack = require('webpack');
+module.exports = {
+    entry: { main: './index.js' },
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: '[name].[chunkhash].js'
+    },
+    module: {
+        rules: [
+            {
+                test: /\.(png|jpe?g|gif|ico|svg)$/i,
+                use: [
+                  'file-loader?name=./images/[name].[ext]',
+                  {
+                    loader: 'image-webpack-loader',
+                    options: {
+                      bypassOnDebug: true, // webpack@1.x
+                      disable: true // webpack@2.x and newer
+                    }
+                  }
+                ]
+              },
+              {
+                test: /\.(eot|ttf|woff|woff2)$/,
+                loader: 'file-loader?name=./fonts/[name].[ext]'
+              },
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: "babel-loader"
+                }
+            },
+            {
+                test: /\.css$/,
+                use:  [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'] // добавили минификацию CSS
+            }
+        ]
+    },
+    plugins: [
+        new webpack.DefinePlugin({
+			'NODE_ENV': JSON.stringify('production')
+		 }),
+        new MiniCssExtractPlugin({ // 
+            filename: 'style.[contenthash].css',
+        }),
+        new HtmlWebpackPlugin({
+            inject: false,
+            hash: true,
+            template: './index.html',
+            filename: 'index.html'
+        }),
+        new WebpackMd5Hash()
+    ]
+};
